+++ src/Modules/Basket/DTO/BasketForListView.php
@@ -12,7 +12,7 @@ class BasketForListView
     public string $description;
     public ?string $picture;
     public int $until;
-    public float $distance;
+    public float $distanceInKm;
     public Profile $creator;
 
     public static function createFromArray(array $data): BasketForListView
@@ -23,7 +23,7 @@ class BasketForListView
         $picture = json_decode($data['picture'] ?? '', true);
         $basket->picture = is_array($picture) ? ($picture[0] ?? null) : $data['picture'];
         $basket->until = $data['until_ts'];
-        $basket->distance = $data['distance'];
+        $basket->distanceInKm = $data['distance_in_km'];
         $basket->creator = new Profile($data, 'fs_');
 
         return $basket;